requirements.txt
appdirs==1.4.0
Django==1.10.5
djangorestframework==3.6.2
drf-base64==0.9.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
